requirements.txt
#ToDo cleanup
pandas
fastapi
databases
pydantic
asyncpg
bcrypt
requests
datetime
pyserial
platformio